+++ example/lib/widgets/icons.dart
@@ -1,22 +1,32 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
 import 'package:flutter/widgets.dart';
 
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: non_constant_identifier_names
+// ignore_for_file: constant_identifier_names
 @immutable
 class _UiIconsData extends IconData {
-  const _UiIconsData(int codePoint)
+  const _UiIconsData(int codePoint, this.name)
       : super(
           codePoint,
           fontFamily: 'UiIcons',
         );
+
+  final String name;
 }
 
 @immutable
 class UiIcons {
   const UiIcons._();
 
-  static const IconData account = _UiIconsData(0xf101);
-  static const IconData arrow_left = _UiIconsData(0xf102);
-  static const IconData arrow_right = _UiIconsData(0xf103);
-  static const IconData collection = _UiIconsData(0xf104);
+  static const account = _UiIconsData(0xf101, 'account');
+  static const arrow_left = _UiIconsData(0xf102, 'arrow_left');
+  static const arrow_right = _UiIconsData(0xf103, 'arrow_right');
+  static const collection = _UiIconsData(0xf104, 'collection');
+
+  static const all = <String, _UiIconsData>{
+    'account': account,
+    'arrow_left': arrow_left,
+    'arrow_right': arrow_right,
+    'collection': collection,
+  };
 }